# Conflicts:
#	AndEngine-GLES2/src/org/andengine/opengl/texture/atlas/source/BaseTextureAtlasSource.java
#	AndEngine/build.gradle
#	AndEngine/src/org/anddev/andengine/entity/text/ChangeableText.java
#	AndEngine/src/org/anddev/andengine/entity/text/Text.java
#	AndEngine/src/org/anddev/andengine/entity/text/TickerText.java
#	AndEngine/src/org/anddev/andengine/opengl/font/Font.java
#	AndEngine/src/org/anddev/andengine/opengl/font/FontFactory.java
#	AndEngine/src/org/anddev/andengine/opengl/font/StrokeFont.java
#	AndEngine/src/org/anddev/andengine/opengl/texture/buffer/TextTextureBuffer.java
#	src/com/edlplan/ui/fragment/SearchBarFragment.kt
#	src/com/osudroid/ui/BannerManager.kt
#	src/com/osudroid/ui/v2/game/FollowPoints.kt
#	src/com/osudroid/ui/v2/hud/GameplayHUD.kt
#	src/com/osudroid/ui/v2/hud/HUDElement.kt
#	src/com/osudroid/ui/v2/hud/editor/HUDElementOverlay.kt
#	src/com/osudroid/ui/v2/hud/editor/HUDElementSelector.kt
#	src/com/osudroid/ui/v2/hud/elements/HUDBarHitErrorMeter.kt
#	src/com/osudroid/ui/v2/hud/elements/HUDColorHitErrorMeter.kt
#	src/com/reco1l/andengine/UIResourceManager.kt
#	src/com/reco1l/andengine/UIScene.kt
#	src/com/reco1l/andengine/buffered/CompoundBuffer.kt
#	src/com/reco1l/andengine/buffered/VertexBuffer.kt
#	src/com/reco1l/andengine/container/UIConstraintContainer.kt
#	src/com/reco1l/andengine/shape/UICircle.kt
#	src/com/reco1l/andengine/text/UIText.kt
#	src/com/reco1l/andengine/ui/UISlider.kt
#	src/ru/nsu/ccfit/zuev/osu/ResourceManager.java

+++ src/ru/nsu/ccfit/zuev/osu/menu/ScoreBoardItem.java
@@ -21,7 +21,7 @@ public class ScoreBoardItem implements Cloneable {
 
 
     public String userName;
-    public int playScore;
+    public long playScore;
     public int scoreId;
 
     /**
@@ -48,7 +48,7 @@ public class ScoreBoardItem implements Cloneable {
 
     public ScoreBoardItem() {}
 
-    public ScoreBoardItem(String userName, int playScore, int maxCombo, float accuracy, boolean isAlive) {
+    public ScoreBoardItem(String userName, long playScore, int maxCombo, float accuracy, boolean isAlive) {
         this.userName = userName;
         this.playScore = playScore;
         this.maxCombo = maxCombo;
@@ -57,7 +57,7 @@ public class ScoreBoardItem implements Cloneable {
     }
 
 
-    public void set(int rankPos, String name, int combo, int score, int id) {
+    public void set(int rankPos, String name, int combo, long score, int id) {
         rank = rankPos;
         userName = name;
         maxCombo = combo;
